+++ kernel/src/sched/proc.zig
@@ -36,7 +36,7 @@ pub const Thread = struct {
     ctx: cpu.Context = std.mem.zeroes(cpu.Context),
     wait_chan: ?*const anyopaque = null,
     wake_tick: u64 = 0,
-    stack_phys: u64,
+    stack_phys: usize,
     proc_node: std.DoublyLinkedList.Node,
     sched_node: std.DoublyLinkedList.Node,
     on_runqueue: bool,
